% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediatorANDoutcome_models.R
\name{data_models}
\alias{data_models}
\title{Generate the dataframe stablishing the fitted models for mediator and the
fitted models for outcome that need to be perform}
\usage{
data_models(
  outcome,
  mediator,
  treatment,
  col_m = "model.m.formula",
  col_y = "model.y.formula"
)
}
\arguments{
\item{outcome}{a vector with the outcomes}

\item{mediator}{a vector with the mediators}

\item{treatment}{a vector with the treatments}

\item{col_m}{a character. Name of the column that will contain the mediator
model formulas in the output dataframe. Default: "model.m.formula".}

\item{col_y}{a character. Name of the column that will contain the outcome
model formulas in the output dataframe. Default: "model.y.formula".}
}
\value{
It returns a dataframe with five columns. The first two columns will
include the outcomes, treatments and mediators tested, respectively. The
last two columns will indicate the fitted models for mediators and outcome
that need to be performed.
}
\description{
Generate the dataframe stablishing the fitted models for mediator and the
fitted models for outcome that need to be perform
}
