% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_medresults.R
\name{formatting_med}
\alias{formatting_med}
\title{Formatting results from causal mediation analysis}
\usage{
formatting_med(mediation.list)
}
\arguments{
\item{mediation.list}{list of lists, with as many sublists as different
outcomes has been tested in the causal mediation analysis.}
}
\value{
lists of lists with the mediate analysis filtered
}
\description{
\code{formatting_med()} extracts the information from the causal
mediation analysis to filter statistically significant mediation results
}
