% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediatorANDtreatment_models.R
\name{generating_models}
\alias{generating_models}
\title{Generating the fitted models for mediators OR outcomes}
\usage{
generating_models(
  column.models,
  model.type,
  data,
  data.models,
  model.m = TRUE,
  outcome = NULL,
  data.split = NULL,
  ...
)
}
\arguments{
\item{column.models}{a character indicating the name of the column containing
the fitted models for mediators OR outcomes.}

\item{model.type}{a function indicating the kind of analysis that will be
performed, taking into account the ones allowed by
\code{\link[mediation]{mediate}}.}

\item{data}{a dataframe with the information to perform the statistical
models.}

\item{data.models}{a dataframe with the column specified in column.models.
This will contain the formulas for the models as characters.}

\item{model.m}{a boolean for choosing if we are going to perform the fitted
models for mediator (TRUE) or outcome (FALSE). Default: TRUE}

\item{outcome}{a string. This will refer to the name of the column that
contains the outcome. Run in the default mode, the function understands
that all analyses share the same outcome. Default: NULL.}

\item{data.split}{a character indicating the column from data to split the
statistical analysis. Default: NULL.}

\item{...}{other arguments that models will need. Some functions, as
\code{\link[stats]{glm}}, requires additional arguments, such as family,
that can be specified here.}
}
\value{
returns a dataframe with a column, named model.M or model.Y,
depending on the fitted models performed
}
\description{
This function generates a dataframe with the fitted models
performed for the mediators or the outcomes (depending on the input
provided). To do so, we will need two objects: * data: a dataframe with the
values to perform the statistical models. * data.models: a dataframe with
the information of the models that need to be performed.
}
