% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hightmediation.R
\name{htmed}
\alias{htmed}
\title{High-throughput causal mediation analysis}
\usage{
htmed(
  data.models,
  column.modelm,
  column.modely,
  treat,
  mediator,
  outcome,
  seed = NULL,
  data.split = NULL,
  ...
)
}
\arguments{
\item{data.models}{a dataframe with, at least, five columns: for the
treatments, mediators, outcomes, fitted models for treatments and
mediators, respectively.}

\item{column.modelm}{a character. Name of the column with the fitted models
for mediators.}

\item{column.modely}{a character. Name of the column with the fitted models
for outcomes.}

\item{treat}{a character. Name of the column that contains the treatments.}

\item{mediator}{a character. Name of the column that contains the mediators.}

\item{outcome}{a character. Name of the column that contains the outcomes.}

\item{seed}{integer to set a seed (for reproducibility). Default: NULL}

\item{data.split}{a character indicating the column from data.models to split
the mediation analysis. Default: NULL.}

\item{...}{other arguments passed to \code{\link[mediation]{mediate}}
function.}
}
\value{
returns a list of lists with the results of mediation for each
combination of outcome, mediator and treatment variables.
}
\description{
\code{hightmed()} allows to perform high-throughput causal mediation
analysis, using the \code{\link[mediation]{mediate}} for each model.
}
