% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediatorANDtreatment_models.R
\name{outcome_models}
\alias{outcome_models}
\title{Generating the fitted models for outcomes}
\usage{
outcome_models(
  column.models,
  model.type,
  data,
  data.models,
  model.name,
  outcome = NULL,
  data.split = NULL,
  ncores = NULL,
  ...
)
}
\arguments{
\item{column.models}{a character indicating the name of the column containing
the fitted models for outcomes OR outcomes.}

\item{model.type}{a function indicating the kind of analysis that will be
performed, taking into account the ones allowed by
\code{\link[mediation]{mediate}}.}

\item{data}{a dataframe with the information to perform the statistical
models.}

\item{data.models}{a dataframe with the column specified in column.models.
This will contain the formulas for the models as characters.}

\item{model.name}{a string referring to the name of the column that will
contain the fitted models for the outcomes.}

\item{outcome}{a string. This will refer to the name of the column that
contains the outcome. Run in the default mode, the function understands
that all analyses share the same outcome. Default: NULL.}

\item{data.split}{a character indicating the column from data to split the
statistical analysis. Default: NULL.}

\item{ncores}{numeric. It refers to the number of cores used during the
mediation analyses. In the default mode, it will detect automatically the
number of cores to choose. Default: NULL.}

\item{...}{other arguments that models will need. Some functions, as
\code{\link[stats]{glm}}, requires additional arguments, such as family,
that can be specified here.}
}
\value{
returns a dataframe with a column, named model.M or model.Y,
depending on the fitted models performed
}
\description{
This function generates a dataframe with the fitted models
performed for the outcomes. To do so, we will need two objects: * data: a
dataframe with the values to perform the statistical models. * data.models:
a dataframe with the information of the models that need to be performed.
}
