% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediatorANDtreatment_models.R
\name{providing_models}
\alias{providing_models}
\title{Merging the fitted models for mediator and outcome}
\usage{
providing_models(model.m, model.y)
}
\arguments{
\item{model.m}{a list with the fitted models for mediator}

\item{model.y}{a list with the fitted models for outcome}
}
\value{
returns a dataframe with two different columns
\itemize{
\item model.M: column with the fitted models for mediator
\item model.Y: column with the fitted models for outcome
}
}
\description{
\code{providing_models()} generates a single dataframe with as many
rows as different models to perform mediation and two columns (one with the
fitted models for mediator and another with fitted models for outcome)
}
