% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visual_htmed}
\alias{visual_htmed}
\title{Visualization of the high-throughput causal mediation analysis}
\usage{
visual_htmed(mediation.list, outcome, ...)
}
\arguments{
\item{mediation.list}{lists of lists with the results of mediation}

\item{outcome}{name of the outcome the user want to visualice}

\item{...}{rest of arguments passed to ggplot}
}
\value{
returns a ggplot object
}
\description{
\code{visual_htmed()} enables to visualize the results of the causal
mediation analyses performed in a single plot
}
