% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediatorANDtreatment_models.v2.R
\name{generating_models}
\alias{generating_models}
\title{Generating the fitted models for mediator OR outcome}
\usage{
generating_models(
  column.models,
  model.type,
  data,
  data.models,
  ncores,
  model.m = TRUE,
  ...
)
}
\arguments{
\item{column.models}{a character indicating the name of the column containing the fitted models for mediator}

\item{model.type}{a function indicating the kind of analysis that will be performed, taking into account the ones allowed by mediate()}

\item{data}{a dataframe with the information to perform the models}

\item{data.models}{a dataframe with the column indicated in column.models}

\item{ncores}{number of ncores to use}

\item{model.m}{Default: TRUE. A boolean for choosing if we are going to perform the fitted models for mediator (TRUE) or outcome (FALSE)}
}
\value{
returns a dataframe with a column, named model.M or model.Y, depending on the fitted models performed
}
\description{
\code{generating_models()} generates a single dataframe with as many
rows as different models to perform mediation and a column (with the
fitted models for mediator OR the fitted models for outcome)
}
