% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{graph_htmed}
\alias{graph_htmed}
\title{Graph summary of the high-throughput causal mediation analysis}
\usage{
graph_htmed(
  mediation.list,
  outcome,
  vertex.label.cex = 0.75,
  vertex.size = 16,
  edge.width = 2,
  edge.arrow.size = 0.5,
  ...
)
}
\arguments{
\item{mediation.list}{lists of lists with the results of mediation}

\item{outcome}{name of the outcome the user want to visualice}

\item{vertex.label.cex}{number indicating the font size of the node}

\item{vertex.size}{number indicating the size of the node}

\item{edge.width}{number indicating the edge width}

\item{edge.arrow.size}{number indicating the arrow size}

\item{...}{rest of arguments passed to igraph}
}
\value{
returns a ggplot object
}
\description{
Graph summary of the high-throughput causal mediation analysis
}
