% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{graph_htmed}
\alias{graph_htmed}
\title{Graph summary of the high-throughput causal mediation analysis}
\usage{
graph_htmed(
  mediation.form,
  outcome,
  pval.column = NULL,
  pval = 0.05,
  prop.med = "Estimate_Prop._Mediated_(average)",
  acme = "Estimate_ACME_(average)",
  treatment = "treatment",
  mediator = "mediator",
  split = NULL
)
}
\arguments{
\item{mediation.form}{lists of lists with the summary of the mediation
analyses}

\item{outcome}{a character indicating the name of the outcome the user wants
to visualice}

\item{pval.column}{a character with the name of the column that contains the
p-values. Default: NULL}

\item{pval}{a number establishing the threshold for the \code{pval.column}}

\item{prop.med}{a character indicating the column with the Prop.Mediated
information. Default: Estimate_Prop.\emph{Mediated}(average).}

\item{acme}{a character indicating the column with the ACME information.
Default: Estimate_ACME_(average).}

\item{treatment}{a character with the name of the column containing the
treatment information. Default: treatment}

\item{mediator}{a character with the name of the column containing the
mediator information. Default: mediator.}

\item{split}{a character indicating the name of the column used for the
split. Default: NULL}
}
\value{
returns a graph with an inner circle (treatments) and an outer circle (mediators).
The width of the edges will indicate the proportion of mediation. The color of the edges refers to the estimator of mediation.
}
\description{
\code{graph_htmed()} enables to visualize the results of the causal
mediation analyses for a specific outcome in a single graph

This function also allows to filter results by p-value, giving the column
that contains this information and the threshold of significance. If not,
the scatterplot will display all results, without taking into account the
significance reported.

In case split argument is provided, as many graphs as different
levels in the split column will be generated.
}
