% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{graph_htmed}
\alias{graph_htmed}
\title{Graph summary of the high-throughput causal mediation analysis}
\usage{
graph_htmed(
  mediation.form,
  outcome,
  pval.column = NULL,
  pval = 0.05,
  prop.med = "Estimate_Prop._Mediated_(average)",
  acme = "Estimate_ACME_(average)",
  treatment = "treatment",
  mediator = "mediator",
  split = NULL,
  ...
)
}
\arguments{
\item{mediation.form}{lists of lists with the results of mediation}

\item{outcome}{name of the outcome the user want to visualice}

\item{...}{rest of arguments passed to igraph}

\item{vertex.label.cex}{number indicating the font size of the node}

\item{vertex.size}{number indicating the size of the node}

\item{edge.width}{number indicating the edge width}

\item{edge.arrow.size}{number indicating the arrow size}
}
\value{
prints out a graph from igraph package
}
\description{
Graph summary of the high-throughput causal mediation analysis
}
