% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visual_htmed}
\alias{visual_htmed}
\title{Scatterplot with high-throughput causal mediation analyses}
\usage{
visual_htmed(
  mediation.form,
  outcome,
  pval.column = NULL,
  pval = 0.05,
  prop.med = "Estimate_Prop._Mediated_(average)",
  acme = "Estimate_ACME_(average)",
  treatment = "treatment",
  mediator = "mediator",
  data.split = NULL
)
}
\arguments{
\item{mediation.form}{lists of lists with the summary of the mediation
analyses}

\item{outcome}{a character indicating the name of the outcome the user wants
to visualice}

\item{pval.column}{a character with the name of the column that contains the
p-values. Default: NULL}

\item{pval}{a number establishing the threshold for the \code{pval.column}}

\item{prop.med}{a character indicating the column with the Prop.Mediated
information. Default: Estimate_Prop.\emph{Mediated}(average).}

\item{acme}{a character indicating the column with the ACME information.
Default: Estimate_ACME_(average).}

\item{treatment}{a character with the name of the column containing the
treatment information. Default: treatment}

\item{mediator}{a character with the name of the column containing the
mediator information. Default: mediator.}

\item{data.split}{a character indicating the name of the column used for the
split. Default: NULL}
}
\value{
returns a scatterplot with the treatments in the X axis and the mediators in the Y axis.
The size of the dots will indicate the proportion of mediation. The color of the dots refers to the estimator of mediation.
}
\description{
\code{visual_htmed()} enables to visualize the results of the causal
mediation analyses for a specific outcome in a single scatterplot.

This function also allows to filter results by p-value, giving the column
that contains this information and the threshold of significance. If not,
the scatterplot will display all results, without taking into account the
significance reported.

In case split argument is provided, as many scatterplots as different
levels in the split column will be generated.
}
